Extract merge pass1 into its own module and tighten size lints
Catalog merge was still a ~2.2k-line unit after the rows/schema split.
Move the pass1 walk (entry resolution, row merge, index patches) into
doltlite_merge_pass1.c, expand the shared merge int header, and wire
main.mk plus the amalgamation.

Layer lints now require the core merge files to exist and cap them at
1500 lines (cmd/constraints at 2000) so a re-monolith cannot slip past
by only checking survivors.

Verified: lint_layers; merge, index-conflict, ignore-corners,
schema-merge, conflicts, cherry-pick, and conflict-rows shell suites.
